Swiperにタイムテーブル用のスライドを追加
